Fix build on ARM architecture

+++ kdeui/widgets/kmessagewidget.cpp
@@ -162,7 +162,7 @@ void KMessageWidgetPrivate::updateSnapShot()
 
 void KMessageWidgetPrivate::slotTimeLineChanged(qreal value)
 {
-    q->setFixedHeight(qMin(value * 2, 1.) * content->height());
+    q->setFixedHeight(qMin(value * 2, qreal(1.0)) * content->height());
     q->update();
 }
